changed preferences tab name to customisation and moved preferences check logic to IdentifierEditorViewModel
