+++ src/main/java/org/jabref/gui/preferences/PreferencesDialogViewModel.java
@@ -14,7 +14,7 @@ import org.jabref.gui.Globals;
 import org.jabref.gui.JabRefFrame;
 import org.jabref.gui.preferences.appearance.AppearanceTab;
 import org.jabref.gui.preferences.citationkeypattern.CitationKeyPatternTab;
-import org.jabref.gui.preferences.doi.DOITab;
+import org.jabref.gui.preferences.customisation.CustomisationTab;
 import org.jabref.gui.preferences.entryeditor.EntryEditorTab;
 import org.jabref.gui.preferences.entryeditortabs.CustomEditorFieldsTab;
 import org.jabref.gui.preferences.exporter.ExportCustomizationTab;
@@ -75,7 +75,7 @@ public class PreferencesDialogViewModel extends AbstractViewModel {
                 new JournalAbbreviationsTab(),
                 new GroupsTab(),
                 new EntryEditorTab(),
-                new DOITab(),
+                new CustomisationTab(),
                 new CustomEditorFieldsTab(),
                 new CitationKeyPatternTab(),
                 new LinkedFilesTab(),
