+++ src/main/java/org/jabref/gui/preferences/customisation/CustomisationTab.fxml
@@ -7,8 +7,8 @@
 <?import javafx.scene.layout.HBox?>
 <fx:root spacing="10.0" type="VBox"
          xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
-         fx:controller="org.jabref.gui.preferences.doi.DOITab">
-    <Label styleClass="titleHeader" text="DOI"/>
+         fx:controller="org.jabref.gui.preferences.customisation.CustomisationTab">
+    <Label styleClass="titleHeader" text="%Customisation"/>
 
     <Label styleClass="sectionHeader" text="%Custom DOI URI"/>
     <HBox alignment="CENTER_LEFT" spacing="10.0">
